boost::log

+++ mkfs.cpp
@@ -2,7 +2,11 @@
 
 int main(int argc, char ** argv)
 {
-	FS * fs = new FS(argv[1], argv[2]);
+	if (argc < 2) {
+		fprintf(stderr, "set db first\n");
+		return -1;
+	}
+	FS * fs = new FS(argv[1]);
 	fs->mkfs();
 	delete fs;
 	return 0;
